Integrate TokenTree
* integrate with OnDiskSA & SASI
* add value merging support to lazy merge iterator (+ fix SASI to avoid reading DK from SSTable)
* Integrate NoOp and Standard tokenizers/filters (+ SA index write refactoring)

+++ src/java/org/apache/cassandra/db/index/search/Term.java
@@ -9,9 +9,9 @@ public class Term
 {
     public final int position;
     public final ByteBuffer value;
-    public final NavigableMap<DecoratedKey, Integer> keys;
+    public final NavigableMap<DecoratedKey, Long> keys;
 
-    public Term(int position, ByteBuffer value, NavigableMap<DecoratedKey, Integer> keys)
+    public Term(int position, ByteBuffer value, NavigableMap<DecoratedKey, Long> keys)
     {
         this.position = position;
         this.value = value;
